refactor(mero-sign): split into library plus thin binary
