+++ tools/mero-sign/src/lib.rs
@@ -1,0 +1,458 @@
+//! mero-sign: library for signing Calimero bundle manifests
+//!
+//! Implements the signing flow:
+//! - Generate Ed25519 keypairs
+//! - Sign manifests using RFC 8785 canonicalization
+//! - Derive did:key signerId from public keys
+
+use std::fs;
+use std::path::Path;
+
+use base64::engine::general_purpose::URL_SAFE_NO_PAD;
+use base64::Engine;
+use ed25519_dalek::{Signature, Signer, SigningKey, Verifier, VerifyingKey};
+use eyre::{bail, Context, Result};
+use serde::{Deserialize, Serialize};
+use sha2::{Digest, Sha256};
+
+/// Multicodec indicator for ed25519-pub (0xed01, varint encoded).
+const ED25519_PUB_MULTICODEC: [u8; 2] = [0xed, 0x01];
+
+/// Well-known development signing key seed.
+///
+/// Derived deterministically: `SHA-256("calimero-dev-signing-key-v1")`.
+/// This key is PUBLIC and provides no security. It exists solely as a
+/// marker for development builds, analogous to Android's `debug.keystore`.
+const DEV_SEED: [u8; 32] = [
+    0x8c, 0xb7, 0xae, 0x25, 0x1b, 0x4d, 0xa5, 0x5a, 0xec, 0xc3, 0x87, 0x5f, 0x4c, 0x0a, 0x71, 0x2c,
+    0xe2, 0xb3, 0x4c, 0xb0, 0x6d, 0x47, 0x71, 0xe7, 0x13, 0xf2, 0xbb, 0x32, 0x89, 0xd0, 0xf7, 0xb4,
+];
+
+/// Key file format
+#[derive(Debug, Serialize, Deserialize)]
+pub struct KeyFile {
+    /// Base64url encoded 32-byte private key seed
+    private_key: String,
+    /// Base64url encoded 32-byte public key
+    public_key: String,
+    /// The derived did:key signerId
+    signer_id: String,
+}
+
+/// Signature object in manifest
+#[derive(Debug, Serialize, Deserialize)]
+struct SignatureObject {
+    algorithm: String,
+    #[serde(rename = "publicKey")]
+    public_key: String,
+    signature: String,
+}
+
+/// Derives a did:key signerId from an Ed25519 public key.
+///
+/// The did:key format for Ed25519 is:
+/// - `did:key:` prefix
+/// - multibase base58btc encoding ('z' prefix)
+/// - multicodec indicator for ed25519-pub (0xed01)
+/// - the 32-byte Ed25519 public key
+pub fn derive_signer_id_did_key(pubkey: &[u8; 32]) -> String {
+    // Construct the multicodec-prefixed key
+    let mut multicodec_key = Vec::with_capacity(2 + 32);
+    multicodec_key.extend_from_slice(&ED25519_PUB_MULTICODEC);
+    multicodec_key.extend_from_slice(pubkey);
+
+    // Encode with base58btc (multibase 'z' prefix)
+    let encoded = bs58::encode(&multicodec_key).into_string();
+
+    format!("did:key:z{encoded}")
+}
+
+/// Canonicalizes a manifest JSON value using RFC 8785 (JCS).
+/// Removes the signature field and all underscore-prefixed transient fields before canonicalization.
+pub fn canonicalize_manifest(manifest_json: &serde_json::Value) -> Result<Vec<u8>> {
+    // Clone and remove the signature field for canonicalization
+    let mut signing_view = manifest_json.clone();
+    if let Some(obj) = signing_view.as_object_mut() {
+        obj.remove("signature");
+        // Remove all underscore-prefixed fields to prevent signature confusion.
+        obj.retain(|k, _| !k.starts_with('_'));
+    }
+
+    // Canonicalize using RFC 8785 JCS
+    let canonical_bytes = serde_json_canonicalizer::to_vec(&signing_view)
+        .context("failed to canonicalize manifest JSON")?;
+
+    Ok(canonical_bytes)
+}
+
+/// Computes the signing payload (SHA-256 hash of canonical manifest bytes).
+pub fn compute_signing_payload(canonical_bytes: &[u8]) -> [u8; 32] {
+    let mut hasher = Sha256::new();
+    hasher.update(canonical_bytes);
+    hasher.finalize().into()
+}
+
+/// Load a signing key from a key file
+pub fn load_signing_key(key_path: &Path) -> Result<SigningKey> {
+    let key_content = fs::read_to_string(key_path)
+        .with_context(|| format!("failed to read key file: {}", key_path.display()))?;
+
+    let key_file: KeyFile = serde_json::from_str(&key_content)
+        .with_context(|| format!("failed to parse key file: {}", key_path.display()))?;
+
+    let private_key_bytes = URL_SAFE_NO_PAD
+        .decode(&key_file.private_key)
+        .context("failed to decode private key from base64url")?;
+
+    if private_key_bytes.len() != 32 {
+        bail!(
+            "invalid private key length: expected 32 bytes, got {}",
+            private_key_bytes.len()
+        );
+    }
+
+    let mut seed = [0u8; 32];
+    seed.copy_from_slice(&private_key_bytes);
+
+    Ok(SigningKey::from_bytes(&seed))
+}
+
+pub fn dev_signing_key() -> SigningKey {
+    SigningKey::from_bytes(&DEV_SEED)
+}
+
+/// Returns the well-known dev signer_id (derived from DEV_SEED at runtime).
+pub fn dev_signer_id() -> String {
+    let key = dev_signing_key();
+    derive_signer_id_did_key(key.verifying_key().as_bytes())
+}
+
+/// Sign a manifest file
+pub fn sign_manifest(manifest_path: &Path, signing_key: &SigningKey, is_dev: bool) -> Result<()> {
+    let verifying_key = signing_key.verifying_key();
+    let signer_id = derive_signer_id_did_key(verifying_key.as_bytes());
+
+    // Read the manifest
+    let manifest_content = fs::read_to_string(manifest_path)
+        .with_context(|| format!("failed to read manifest: {}", manifest_path.display()))?;
+
+    let mut manifest: serde_json::Value = serde_json::from_str(&manifest_content)
+        .with_context(|| format!("failed to parse manifest: {}", manifest_path.display()))?;
+
+    // Add signerId to the manifest
+    if let Some(obj) = manifest.as_object_mut() {
+        obj.insert(
+            "signerId".to_string(),
+            serde_json::Value::String(signer_id.clone()),
+        );
+        // Also add minRuntimeVersion if not present
+        if !obj.contains_key("minRuntimeVersion") {
+            obj.insert(
+                "minRuntimeVersion".to_string(),
+                serde_json::Value::String("0.1.0".to_string()),
+            );
+        }
+    }
+
+    // Canonicalize the manifest (without signature)
+    let canonical_bytes = canonicalize_manifest(&manifest)?;
+
+    // Compute the signing payload
+    let signing_payload = compute_signing_payload(&canonical_bytes);
+
+    // Sign the payload
+    let signature = signing_key.sign(&signing_payload);
+
+    // Encode public key and signature as base64url
+    let public_key_b64 = URL_SAFE_NO_PAD.encode(verifying_key.as_bytes());
+    let signature_b64 = URL_SAFE_NO_PAD.encode(signature.to_bytes());
+
+    // Add the signature object to the manifest
+    let signature_obj = SignatureObject {
+        algorithm: "ed25519".to_string(),
+        public_key: public_key_b64,
+        signature: signature_b64,
+    };
+
+    if let Some(obj) = manifest.as_object_mut() {
+        obj.insert(
+            "signature".to_string(),
+            serde_json::to_value(&signature_obj)?,
+        );
+    }
+
+    // Write the signed manifest back
+    let signed_manifest = serde_json::to_string_pretty(&manifest)?;
+    fs::write(manifest_path, signed_manifest).with_context(|| {
+        format!(
+            "failed to write signed manifest: {}",
+            manifest_path.display()
+        )
+    })?;
+
+    if is_dev {
+        eprintln!(
+            "\u{26a0}  Signed with DEVELOPMENT key. This bundle cannot be published to the registry."
+        );
+    } else {
+        eprintln!("Signed manifest: {}", manifest_path.display());
+    }
+    eprintln!("  signerId: {signer_id}");
+
+    Ok(())
+}
+
+/// Verify a signed manifest end to end: re-canonicalize (RFC 8785, signature and
+/// underscore fields excluded), hash, and check the ed25519 signature against the
+/// embedded public key, then confirm `signerId` derives from that same key.
+///
+/// Returns `Ok(true)` when the signature verifies, `Ok(false)` when it is present
+/// but invalid, and `Err` when the manifest lacks a well-formed `signature`
+/// object. Mirrors the node's `verify_manifest_signature` gate.
+pub fn verify_manifest(manifest: &serde_json::Value) -> Result<bool> {
+    let signature_obj = manifest
+        .get("signature")
+        .ok_or_else(|| eyre::eyre!("manifest is missing the 'signature' field"))?;
+
+    let algorithm = signature_obj
+        .get("algorithm")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| eyre::eyre!("signature missing 'algorithm' field"))?;
+    if algorithm != "ed25519" {
+        bail!("unsupported signature algorithm: '{algorithm}', expected 'ed25519'");
+    }
+
+    let public_key_b64 = signature_obj
+        .get("publicKey")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| eyre::eyre!("signature missing 'publicKey' field"))?;
+    let signature_b64 = signature_obj
+        .get("signature")
+        .and_then(|v| v.as_str())
+        .ok_or_else(|| eyre::eyre!("signature missing 'signature' field"))?;
+
+    let public_key_bytes: [u8; 32] = URL_SAFE_NO_PAD
+        .decode(public_key_b64)
+        .context("failed to decode publicKey from base64url")?
+        .try_into()
+        .map_err(|_| eyre::eyre!("publicKey is not 32 bytes"))?;
+    let signature_bytes = URL_SAFE_NO_PAD
+        .decode(signature_b64)
+        .context("failed to decode signature from base64url")?;
+
+    let verifying_key = VerifyingKey::from_bytes(&public_key_bytes)
+        .context("publicKey is not a valid ed25519 point")?;
+    let signature = Signature::from_slice(&signature_bytes)
+        .context("signature is not a valid ed25519 signature")?;
+
+    let canonical_bytes = canonicalize_manifest(manifest)?;
+    let signing_payload = compute_signing_payload(&canonical_bytes);
+
+    if verifying_key.verify(&signing_payload, &signature).is_err() {
+        return Ok(false);
+    }
+
+    // A valid signature over a mismatched signerId is still a forgery vector.
+    if let Some(declared) = manifest.get("signerId").and_then(|v| v.as_str()) {
+        if declared != derive_signer_id_did_key(&public_key_bytes) {
+            return Ok(false);
+        }
+    }
+
+    Ok(true)
+}
+
+/// Generate a new Ed25519 keypair
+pub fn generate_key(output_path: &Path) -> Result<()> {
+    let mut rng = rand::thread_rng();
+    let signing_key = SigningKey::generate(&mut rng);
+    let verifying_key: VerifyingKey = signing_key.verifying_key();
+
+    let private_key_b64 = URL_SAFE_NO_PAD.encode(signing_key.to_bytes());
+    let public_key_b64 = URL_SAFE_NO_PAD.encode(verifying_key.as_bytes());
+    let signer_id = derive_signer_id_did_key(verifying_key.as_bytes());
+
+    let key_file = KeyFile {
+        private_key: private_key_b64,
+        public_key: public_key_b64,
+        signer_id: signer_id.clone(),
+    };
+
+    let key_json = serde_json::to_string_pretty(&key_file)?;
+    fs::write(output_path, key_json)
+        .with_context(|| format!("failed to write key file: {}", output_path.display()))?;
+
+    eprintln!("Generated new keypair: {}", output_path.display());
+    eprintln!("  signerId: {signer_id}");
+
+    Ok(())
+}
+
+/// Derive the signerId from a key file
+pub fn derive_signer_id(key_path: &Path) -> Result<()> {
+    let key_content = fs::read_to_string(key_path)
+        .with_context(|| format!("failed to read key file: {}", key_path.display()))?;
+
+    let key_file: KeyFile = serde_json::from_str(&key_content)
+        .with_context(|| format!("failed to parse key file: {}", key_path.display()))?;
+
+    println!("{}", key_file.signer_id);
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use ed25519_dalek::Verifier;
+    use serde_json::json;
+
+    use super::*;
+
+    fn base64_decode_url(s: &str) -> Vec<u8> {
+        URL_SAFE_NO_PAD.decode(s).unwrap()
+    }
+
+    #[test]
+    fn canonicalize_drops_signature_and_private_fields() {
+        let m = json!({"b": 1, "a": 2, "signature": {"x": 1}, "_local": true});
+        let canon = canonicalize_manifest(&m).unwrap();
+        let s = String::from_utf8(canon).unwrap();
+        assert_eq!(s, r#"{"a":2,"b":1}"#);
+    }
+
+    #[test]
+    fn dev_sign_round_trip() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("manifest.json");
+        std::fs::write(
+            &path,
+            r#"{"package":"com.example.app","appVersion":"1.0.0"}"#,
+        )
+        .unwrap();
+
+        let key = dev_signing_key();
+        sign_manifest(&path, &key, true).unwrap();
+
+        let signed: serde_json::Value =
+            serde_json::from_str(&std::fs::read_to_string(&path).unwrap()).unwrap();
+        assert_eq!(signed["signature"]["algorithm"], "ed25519");
+        assert_eq!(signed["signerId"], dev_signer_id());
+
+        // signature must verify over the SHA-256 digest of the re-canonicalized manifest
+        let canon = canonicalize_manifest(&signed).unwrap();
+        let payload = compute_signing_payload(&canon);
+        let sig_bytes = base64_decode_url(signed["signature"]["signature"].as_str().unwrap());
+        let sig = ed25519_dalek::Signature::from_slice(&sig_bytes).unwrap();
+        key.verifying_key().verify(&payload, &sig).unwrap();
+    }
+
+    #[test]
+    fn verify_manifest_accepts_valid_and_rejects_tampered() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("manifest.json");
+        std::fs::write(
+            &path,
+            r#"{"package":"com.example.app","appVersion":"1.0.0","wasm":{"path":"app.wasm","size":10}}"#,
+        )
+        .unwrap();
+
+        sign_manifest(&path, &dev_signing_key(), true).unwrap();
+        let signed: serde_json::Value =
+            serde_json::from_str(&std::fs::read_to_string(&path).unwrap()).unwrap();
+        assert!(verify_manifest(&signed).unwrap());
+
+        // Mutating a signed field must invalidate the signature.
+        let mut tampered = signed.clone();
+        tampered["appVersion"] = json!("9.9.9");
+        assert!(!verify_manifest(&tampered).unwrap());
+
+        // A signerId that no longer derives from the public key is a forgery.
+        let mut spoofed = signed;
+        spoofed["signerId"] = json!("did:key:zSpoofed");
+        assert!(!verify_manifest(&spoofed).unwrap());
+
+        // No signature object at all is a hard error, not a silent `false`.
+        assert!(verify_manifest(&json!({"package": "x"})).is_err());
+    }
+
+    #[test]
+    fn test_derive_signer_id_did_key() {
+        let pubkey: [u8; 32] = [
+            0x3b, 0x6a, 0x27, 0xbc, 0xce, 0xb6, 0xa4, 0x2d, 0x62, 0xa3, 0xa8, 0xd0, 0x2a, 0x6f,
+            0x0d, 0x73, 0x65, 0x32, 0x15, 0x77, 0x1d, 0xe2, 0x43, 0xa6, 0x3a, 0xc0, 0x48, 0xa1,
+            0x8b, 0x59, 0xda, 0x29,
+        ];
+
+        let signer_id = derive_signer_id_did_key(&pubkey);
+
+        assert!(signer_id.starts_with("did:key:z"));
+        assert!(signer_id.len() > 10);
+    }
+
+    #[test]
+    fn test_generate_and_sign() {
+        let dir = tempfile::tempdir().unwrap();
+        let key_path = dir.path().join("test-key.json");
+        let manifest_path = dir.path().join("manifest.json");
+
+        generate_key(&key_path).unwrap();
+
+        let manifest = json!({
+            "version": "1.0",
+            "package": "com.test.app",
+            "appVersion": "1.0.0",
+            "wasm": {
+                "path": "app.wasm",
+                "size": 1024,
+                "hash": null
+            }
+        });
+        std::fs::write(
+            &manifest_path,
+            serde_json::to_string_pretty(&manifest).unwrap(),
+        )
+        .unwrap();
+
+        let signing_key = load_signing_key(&key_path).unwrap();
+        sign_manifest(&manifest_path, &signing_key, false).unwrap();
+
+        let signed_content = std::fs::read_to_string(&manifest_path).unwrap();
+        let signed_manifest: serde_json::Value = serde_json::from_str(&signed_content).unwrap();
+
+        assert!(signed_manifest.get("signerId").is_some());
+        assert!(signed_manifest.get("signature").is_some());
+        assert!(signed_manifest.get("minRuntimeVersion").is_some());
+
+        let signature = signed_manifest.get("signature").unwrap();
+        assert_eq!(
+            signature.get("algorithm").unwrap().as_str().unwrap(),
+            "ed25519"
+        );
+        assert!(signature.get("publicKey").is_some());
+        assert!(signature.get("signature").is_some());
+    }
+
+    #[test]
+    fn test_dev_key_deterministic() {
+        let key = dev_signing_key();
+        let signer_id = derive_signer_id_did_key(key.verifying_key().as_bytes());
+
+        assert!(signer_id.starts_with("did:key:z"));
+
+        // Must be deterministic across runs
+        let key2 = dev_signing_key();
+        let signer_id2 = derive_signer_id_did_key(key2.verifying_key().as_bytes());
+        assert_eq!(signer_id, signer_id2);
+    }
+
+    #[test]
+    fn test_dev_signer_id_is_stable() {
+        // This is the canonical dev signer_id. If this changes, all dev
+        // ApplicationIds change and existing dev contexts break.
+        let id = dev_signer_id();
+        assert_eq!(
+            id,
+            "did:key:z6MknF3p5L5FDHJQ7FREUapuX4Wmp4MtF6WrHYaXS2B3eZQd"
+        );
+    }
+}
